amazon-s3 Buckect Update

+++ app/Http/Controllers/HomeBladeController.php
@@ -176,11 +176,17 @@ class HomeBladeController extends Controller
 
            if(strstr($mime, "image/")){
 
-             //Delete the Old IMAGE AVATAR from  avatar Folder (Save Space)
-             $imagePath = explode('/', $user_id->avatar);
-             dd($imagePath);
-             $avatarName = $imagePath[5];
-             Storage::disk('s3')->delete('Uploads/avatars/'.$avatarName);
+
+             if ($user_id->avatar !=) {
+               
+               //Delete the Old IMAGE AVATAR from  avatar Folder (Save Space)
+               $imagePath = explode('/', $user_id->avatar);
+               dd($imagePath);
+               $avatarName = $imagePath[5];
+               Storage::disk('s3')->delete('Uploads/avatars/'.$avatarName);
+
+             }
+
 
              //store another avatar
              $user_image = request()->file('avatar');
